Check if APIs present in webhook delete validation (#543)
The delete validation check in the validatingwebhook was exceeding
the default 10 second timeout, preventing deletes. This seemed to
occur when the APIs called don't exist on the server because the
installation hasn't completed. Checking for the APIs presence on
the server with the client-go discovery API before making a list
request performs much faster and prevents webhook timeouts.