Validate deployment labels are updated (#650)
The ocm-antiaffinity-selector label added in 2.2 is not properly applied
to deployments or their pod templates. In 2.3 the services were updated to
set this label in their selector while it is still missing from the pods. This
change will remove the antiaffinity label from the service selector and add it
to the pod template if it is missing.

Signed-off-by: Jakob Gray <20209054+JakobGray@users.noreply.github.com>