Move the virt policies to the global-set namespace
The globalset namespace includes binding to work with all clusters so
it can work well with out of the box policies.  The hub templating
RBAC doesn't seem to work well so there may be a bug limiting the
serviceaccount access, but this should provide the framework for
what is trying to be created.

Signed-off-by: Gus Parvin <gparvin@redhat.com>